Move a couple of EnabledDynamicFieldForm.js functions into their own module
